@@ -1,11 +1,7 @@
 package org.open4goods.services;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
 import java.util.Map.Entry;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
 import org.open4goods.config.yml.attributes.AttributeConfig;
@@ -138,7 +134,7 @@ public class SearchService {
 
 		VerticalSearchResponse vsr = new VerticalSearchResponse();
 
-		List<AttributeConfig> customAttrFilters = vertical.verticalFilters().stream().filter(e-> e!= null).collect(Collectors.toList());
+		List<AttributeConfig> customAttrFilters = vertical.verticalFilters().stream().filter(Objects::nonNull).toList();
 
 
 		Criteria criterias = new Criteria("vertical").is(vertical.getId())
@@ -352,7 +348,7 @@ public class SearchService {
 				vsr.getCustomFilters().get(attrConfig).sort((o1, o2) -> o2.getCount().compareTo(o1.getCount()));
 			}
 			else {
-				vsr.getCustomFilters().get(attrConfig).sort((o1, o2) -> o1.getText().compareTo(o2.getText()));
+				vsr.getCustomFilters().get(attrConfig).sort(Comparator.comparing(VerticalFilterTerm::getText));
 			}
 
 			// TODO : add missing
