@@ -1,32 +1,21 @@
 
 package org.open4goods.api.services;
 
-import java.util.ArrayList;
-import java.util.List;
-
+import jakarta.annotation.PreDestroy;
 import org.open4goods.aggregation.AbstractAggregationService;
 import org.open4goods.aggregation.aggregator.BatchedAggregator;
 import org.open4goods.aggregation.services.aggregation.batch.VerticalBatchedAggregationService;
-import org.open4goods.aggregation.services.aggregation.batch.scores.Attribute2ScoreAggregationService;
-import org.open4goods.aggregation.services.aggregation.batch.scores.Brand2ScoreAggregationService;
-import org.open4goods.aggregation.services.aggregation.batch.scores.CleanScoreAggregationService;
-import org.open4goods.aggregation.services.aggregation.batch.scores.DataCompletion2ScoreAggregationService;
-import org.open4goods.aggregation.services.aggregation.batch.scores.EcoScoreAggregationService;
+import org.open4goods.aggregation.services.aggregation.batch.scores.*;
 import org.open4goods.api.config.yml.ApiProperties;
 import org.open4goods.config.yml.ui.VerticalConfig;
 import org.open4goods.dao.ProductRepository;
-import org.open4goods.services.BarcodeValidationService;
-import org.open4goods.services.BrandService;
-import org.open4goods.services.DataSourceConfigService;
-import org.open4goods.services.EvaluationService;
-import org.open4goods.services.Gs1PrefixService;
-import org.open4goods.services.StandardiserService;
-import org.open4goods.services.VerticalsConfigService;
+import org.open4goods.services.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.config.AutowireCapableBeanFactory;
 
-import jakarta.annotation.PreDestroy;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * This service is in charge of building Product in realtime mode
