@@ -2,6 +2,7 @@ package org.open4goods.ui.config;
 
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadFactory;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,6 +21,6 @@ public class SchedulingConfig implements SchedulingConfigurer {
 
 	@Bean(destroyMethod = "shutdown")
 	public Executor taskExecutor() {
-		return Executors.newScheduledThreadPool(5);
+		return Executors.newScheduledThreadPool(5, Thread.ofVirtual().factory());
 	}
 }
