@@ -6,7 +6,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -24,7 +23,11 @@ import org.springframework.security.web.SecurityFilterChain;
  */
 public class WebConfig {
 
-	private @Autowired AuthenticationProvider  authProvider;
+	private final AuthenticationProvider  authProvider;
+
+	public WebConfig(AuthenticationProvider authProvider) {
+		this.authProvider = authProvider;
+	}
 
 
 	@Bean
